core module from rust
